+++ geotools/src/main/scala/geotrellis/geotools/SimpleFeatureToFeature.scala
@@ -18,7 +18,7 @@ package geotrellis.geotools
 
 import geotrellis.vector._
 
-import com.vividsolutions.jts.{geom => jts}
+import org.locationtech.jts.{geom => jts}
 import org.opengis.feature.simple.SimpleFeature
 
 import scala.collection._