+++ raster/src/main/scala/geotrellis/raster/triangulation/DelaunayRasterizer.scala
@@ -1,6 +1,6 @@
 package geotrellis.raster.triangulation
 
-import com.vividsolutions.jts.geom.Coordinate
+import org.locationtech.jts.geom.Coordinate
 import geotrellis.raster._
 import geotrellis.vector._
 import geotrellis.vector.mesh.{HalfEdgeTable, IndexedPointSet}