+++ vector/src/main/scala/geotrellis/vector/Dimension.scala
@@ -16,7 +16,7 @@
 
 package geotrellis.vector
 
-import com.vividsolutions.jts.{geom => jts}
+import org.locationtech.jts.{geom => jts}
 
 sealed trait Dimensions {
   private[vector] val jtsGeom: jts.Geometry