+++ vector/src/main/scala/geotrellis/vector/MultiGeometry.scala
@@ -16,7 +16,7 @@
 
 package geotrellis.vector
 
-import com.vividsolutions.jts.{geom => jts}
+import org.locationtech.jts.{geom => jts}
 
 trait MultiGeometry extends Geometry {
   /** Computes a area containing these geometries and buffered by size d. */