+++ vector/src/main/scala/geotrellis/vector/io/WKB/WKB.scala
@@ -19,7 +19,7 @@ package geotrellis.vector.io.wkb
 import geotrellis.util.LazyLogging
 import geotrellis.vector._
 
-import com.vividsolutions.jts.io.WKBReader
+import org.locationtech.jts.io.WKBReader
 
 
 /** A thread-safe wrapper for the [https://en.wikipedia.org/wiki/Well-known_text#Well-known_binary WKB]