+++ vector/src/main/scala/geotrellis/vector/triangulation/DelaunayStitcher.scala
@@ -1,6 +1,6 @@
 package geotrellis.vector.triangulation
 
-import com.vividsolutions.jts.geom.Coordinate
+import org.locationtech.jts.geom.Coordinate
 import geotrellis.vector.{Line, MultiLine, Point}
 import geotrellis.vector.RobustPredicates.{LEFTOF, ON, RIGHTOF}
 import geotrellis.vector.mesh.{HalfEdgeTable, IndexedPointSet}